requirements.txt
Package              Version
-------------------- -----------
alembic              1.6.5
annotated-doc        0.0.4
annotated-types      0.7.0
anyio                4.12.1
bcrypt               3.2.2
certifi              2026.1.4
cffi                 2.0.0
click                8.3.1
colorama             0.4.6
cryptography         46.0.3
dnspython            2.8.0
ecdsa                0.19.1
email-validator      2.3.0
fastapi              0.128.0
fastapi-cli          0.0.20
fastapi-cloud-cli    0.11.0
fastar               0.8.0
greenlet             3.3.0
h11                  0.16.0
httpcore             1.0.9
httptools            0.7.1
httpx                0.28.1
idna                 3.11
itsdangerous         2.2.0
Jinja2               3.1.6
Mako                 1.3.10
markdown-it-py       4.0.0
MarkupSafe           3.0.3
mdurl                0.1.2
orjson               3.11.5
passlib              1.7.4
pip                  25.3
psycopg2             2.9.11
pyasn1               0.6.2
pycparser            3.0
pydantic             2.12.5
pydantic_core        2.41.5
pydantic-extra-types 2.11.0
pydantic-settings    2.12.0
Pygments             2.19.2
python-dateutil      2.9.0.post0
python-dotenv        1.2.1
python-editor        1.0.4
python-jose          3.5.0
python-multipart     0.0.21
PyYAML               6.0.3
rich                 14.2.0
rich-toolkit         0.17.1
rignore              0.7.6
rsa                  4.9.1
sentry-sdk           2.50.0
shellingham          1.5.4
six                  1.17.0
SQLAlchemy           1.4.45
starlette            0.50.0
typer                0.21.1
typing_extensions    4.15.0
typing-inspection    0.4.2
tzdata               2025.3
ujson                5.11.0
urllib3              2.6.3
uvicorn              0.40.0
watchfiles           1.1.1
websockets           16.0
